+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 3.4.0 - 2023-07-14
+
+### Added
+
+- Add agent to watch `proofs` directory to trigger tests
+
+### Changed
+
+- `fixtures` and `properties` watcher are now started
+
+### Fixed
+
+- Fix watching `fixtures` and `properties` directories
+- Fix trying to watch the `tests` directory when it doesn't exist
+
 ## 3.3.0 - 2023-07-08
 
 ### Added